+++ source/DecisionTrail.cpp
@@ -5,34 +5,34 @@ namespace sat_solver {
 
     static constexpr std::size_t EmptyIndex = static_cast<std::size_t>(-1);
 
-    DecisionTrail::Entry::Entry(Literal::Int variable, VariableAssignment assn, Reason reason, std::size_t level)
+    DecisionTrail::Entry::Entry(Literal::UInt variable, VariableAssignment assn, Reason reason, std::size_t level)
         : variable{variable}, assignment{assn}, reason{reason}, level{level} {}
 
     DecisionTrail::DecisionTrail(std::size_t num_of_variables)
         : var_index(num_of_variables, EmptyIndex) {}
 
-    void DecisionTrail::Decision(Literal::Int variable, VariableAssignment assn) {
+    void DecisionTrail::Decision(Literal::UInt variable, VariableAssignment assn) {
         this->var_index[variable - 1] = this->trail.size();
         this->trail.emplace_back(variable, assn, ReasonDecision, ++this->level);
     }
 
-    void DecisionTrail::Propagation(Literal::Int variable, VariableAssignment assn) {
+    void DecisionTrail::Propagation(Literal::UInt variable, VariableAssignment assn) {
         this->var_index[variable - 1] = this->trail.size();
         this->trail.emplace_back(variable, assn, ReasonPropagation, this->level);
     }
 
-    void DecisionTrail::Propagation(Literal::Int variable, VariableAssignment assn, std::size_t reason) {
+    void DecisionTrail::Propagation(Literal::UInt variable, VariableAssignment assn, std::size_t reason) {
         this->var_index[variable - 1] = this->trail.size();
         this->trail.emplace_back(variable, assn, static_cast<Reason>(reason), this->level);
     }
 
-    void DecisionTrail::Assumption(Literal::Int variable, VariableAssignment assn) {
+    void DecisionTrail::Assumption(Literal::UInt variable, VariableAssignment assn) {
         this->var_index[variable - 1] = this->trail.size();
         this->trail.emplace_back(variable, assn, ReasonAssumption, ++this->level);
     }
 
     const DecisionTrail::Entry *DecisionTrail::Top() {
-        while (!this->trail.empty() && static_cast<std::size_t>(this->trail.back().variable) > this->var_index.size()) {
+        while (!this->trail.empty() && this->trail.back().variable > this->var_index.size()) {
             this->trail.pop_back();
         }
         if (!this->trail.empty()) {
@@ -43,7 +43,7 @@ namespace sat_solver {
     }
 
     void DecisionTrail::Pop() {
-        while (!this->trail.empty() && static_cast<std::size_t>(this->trail.back().variable) > this->var_index.size()) {
+        while (!this->trail.empty() && this->trail.back().variable > this->var_index.size()) {
             this->trail.pop_back();
         }
         if (!this->trail.empty()) {
@@ -72,8 +72,8 @@ namespace sat_solver {
         }
     }
 
-    const DecisionTrail::Entry *DecisionTrail::Find(Literal::Int variable) const {
-        if (variable - 1 >= static_cast<std::int64_t>(this->var_index.size())) {
+    const DecisionTrail::Entry *DecisionTrail::Find(Literal::UInt variable) const {
+        if (variable - 1 >= this->var_index.size()) {
             return nullptr;
         }
         auto index = this->var_index[variable - 1];
